refactor : Input 컴포넌트 파일 common폴더로 수정
